requirements.txt
pyserial>=3.5
pyyaml>=6.0
asyncio-mqtt>=0.16.0  # For later MQTT integration
aiosqlite>=0.19.0     # For later database